tests/wpt/web-platform-tests/css/css-color/parsing/color-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Module Level 3: getComputedStyle().color</title>
<link rel="help" href="https://drafts.csswg.org/css-color/#propdef-color">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-color-function-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-lab-lch">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-oklab-oklch">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<style>
  #container {
    color: rgb(255, 0, 0);
  }
</style>
<script>
test_computed_value("color", "currentcolor", "rgb(255, 0, 0)");
test_computed_value("color", "transparent", "rgba(0, 0, 0, 0)");
test_computed_value("color", "red", "rgb(255, 0, 0)");
test_computed_value("color", "magenta", "rgb(255, 0, 255)");
test_computed_value("color", "#234", "rgb(34, 51, 68)");
test_computed_value("color", "#FEDCBA", "rgb(254, 220, 186)");
test_computed_value("color", "rgb(2, 3, 4)");
test_computed_value("color", "rgb(100%, 0%, 0%)", "rgb(255, 0, 0)");
test_computed_value("color", "rgba(2, 3, 4, 0.5)");
test_computed_value("color", "rgba(2, 3, 4, 50%)", "rgba(2, 3, 4, 0.5)");
test_computed_value("color", "hsl(120, 100%, 50%)", "rgb(0, 255, 0)");
test_computed_value("color", "hsla(120, 100%, 50%, 0.25)", "rgba(0, 255, 0, 0.25)");
test_computed_value("color", "rgb(-2, 3, 4)", "rgb(0, 3, 4)");
test_computed_value("color", "rgb(100, 200, 300)", "rgb(100, 200, 255)");
test_computed_value("color", "rgb(20, 10, 0, -10)", "rgba(20, 10, 0, 0)");
test_computed_value("color", "rgb(100%, 200%, 300%)", "rgb(255, 255, 255)");

test_computed_value("color", "rgb(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "rgb(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "rgb(128 none none)", "rgb(128, 0, 0)");
test_computed_value("color", "rgb(128 none none / none)", "rgba(128, 0, 0, 0)");
test_computed_value("color", "rgb(none none none / .5)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgb(20% none none)", "rgb(51, 0, 0)");
test_computed_value("color", "rgb(20% none none / none)", "rgba(51, 0, 0, 0)");
test_computed_value("color", "rgb(none none none / 50%)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgba(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "rgba(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "rgba(128 none none)", "rgb(128, 0, 0)");
test_computed_value("color", "rgba(128 none none / none)", "rgba(128, 0, 0, 0)");
test_computed_value("color", "rgba(none none none / .5)", "rgba(0, 0, 0, 0.5)");
test_computed_value("color", "rgba(20% none none)", "rgb(51, 0, 0)");
test_computed_value("color", "rgba(20% none none / none)", "rgba(51, 0, 0, 0)");
test_computed_value("color", "rgba(none none none / 50%)", "rgba(0, 0, 0, 0.5)");

test_computed_value("color", "hsl(120 30% 50%)", "rgb(89, 166, 89)");
test_computed_value("color", "hsl(120 30% 50% / 0.5)", "rgba(89, 166, 89, 0.5)");
test_computed_value("color", "hsl(120 30% 50% / 50%)", "rgba(89, 166, 89, 0.5)");
test_computed_value("color", "hsl(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(0 0% 0%)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "hsl(0 0% 0% / 0)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "hsla(none none none)", "rgb(0, 0, 0)");
test_computed_value("color", "hsla(0 0% 0%)", "rgb(0, 0, 0)");
test_computed_value("color", "hsla(none none none / none)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "hsla(0 0% 0% / 0)", "rgba(0, 0, 0, 0)");
test_computed_value("color", "hsl(120 none none)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(120 0% 0%)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(120 80% none)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(120 80% 0%)", "rgb(0, 0, 0)");
test_computed_value("color", "hsl(120 none 50%)", "rgb(128, 128, 128)");
test_computed_value("color", "hsl(120 0% 50%)", "rgb(128, 128, 128)");
test_computed_value("color", "hsl(120 100% 50% / none)", "rgba(0, 255, 0, 0)");
test_computed_value("color", "hsl(120 100% 50% / 0)", "rgba(0, 255, 0, 0)");
test_computed_value("color", "hsl(120 100% 50% / 0%)", "rgba(0, 255, 0, 0)");
test_computed_value("color", "hsl(none 100% 50%)", "rgb(255, 0, 0)");
test_computed_value("color", "hsl(0 100% 50%)", "rgb(255, 0, 0)");

test_computed_value("color", "hwb(120 30% 50%)", "rgb(77, 128, 77)");
test_computed_value("color", "hwb(120 30% 50% / 0.5)", "rgba(77, 128, 77, 0.5)");
test_computed_value("color", "hwb(120 30% 50% / 50%)", "rgba(77, 128, 77, 0.5)");
test_computed_value("color", "hwb(none none none)", "rgb(255, 0, 0)");
test_computed_value("color", "hwb(0 0% 0%)", "rgb(255, 0, 0)");
test_computed_value("color", "hwb(none none none / none)", "rgba(255, 0, 0, 0)");
test_computed_value("color", "hwb(0 0% 0% / 0)", "rgba(255, 0, 0, 0)");
test_computed_value("color", "hwb(120 none none)", "rgb(0, 255, 0)");
test_computed_value("color", "hwb(120 0% 0%)", "rgb(0, 255, 0)");
test_computed_value("color", "hwb(120 80% none)", "rgb(204, 255, 204)");
test_computed_value("color", "hwb(120 80% 0%)", "rgb(204, 255, 204)");
test_computed_value("color", "hwb(120 none 50%)", "rgb(0, 128, 0)");
test_computed_value("color", "hwb(120 0% 50%)", "rgb(0, 128, 0)");
test_computed_value("color", "hwb(120 30% 50% / none)", "rgba(77, 128, 77, 0)");
test_computed_value("color", "hwb(120 30% 50% / 0)", "rgba(77, 128, 77, 0)");
test_computed_value("color", "hwb(120 30% 50% / 0%)", "rgba(77, 128, 77, 0)");
test_computed_value("color", "hwb(none 100% 50% / none)", "rgba(170, 170, 170, 0)");
test_computed_value("color", "hwb(0 100% 50% / 0)", "rgba(170, 170, 170, 0)");

for (const colorSpace of [ "srgb", "srgb-linear", "a98-rgb", "rec2020", "prophoto-rgb" ]) {
    test_computed_value("color", `color(${colorSpace} 0% 0% 0%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 10% 10% 10%)`, `color(${colorSpace} 0.1 0.1 0.1)`);
    test_computed_value("color", `color(${colorSpace} .2 .2 25%)`, `color(${colorSpace} 0.2 0.2 0.25)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 1)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0% 0 0 / 0.5)`, `color(${colorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 20% 0 10/0.5)`, `color(${colorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 20% 0 10/50%)`, `color(${colorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 400% 0 10/50%)`, `color(${colorSpace} 4 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 50% -160 160)`, `color(${colorSpace} 0.5 -160 160)`);
    test_computed_value("color", `color(${colorSpace} 50% -200 200)`, `color(${colorSpace} 0.5 -200 200)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / -10%)`, `color(${colorSpace} 0 0 0 / 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 110%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 300%)`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 50% -200)`, `color(${colorSpace} 0.5 -200 0)`);
    test_computed_value("color", `color(${colorSpace} 50%)`, `color(${colorSpace} 0.5 0 0)`);
    test_computed_value("color", `color(${colorSpace})`, `color(${colorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 50% -200 / 0.5)`, `color(${colorSpace} 0.5 -200 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 50% / 0.5)`, `color(${colorSpace} 0.5 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} / 0.5)`, `color(${colorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 200 200 200)`, `color(${colorSpace} 200 200 200)`);
    test_computed_value("color", `color(${colorSpace} 200 200 200 / 200)`, `color(${colorSpace} 200 200 200)`);
    test_computed_value("color", `color(${colorSpace} -200 -200 -200)`, `color(${colorSpace} -200 -200 -200)`);
    test_computed_value("color", `color(${colorSpace} -200 -200 -200 / -200)`, `color(${colorSpace} -200 -200 -200 / 0)`);
    test_computed_value("color", `color(${colorSpace} 200% 200% 200%)`, `color(${colorSpace} 2 2 2)`);
    test_computed_value("color", `color(${colorSpace} 200% 200% 200% / 200%)`, `color(${colorSpace} 2 2 2)`);
    test_computed_value("color", `color(${colorSpace} -200% -200% -200% / -200%)`, `color(${colorSpace} -2 -2 -2 / 0)`);
    test_computed_value("color", `color(${colorSpace} calc(0.5 + 1) calc(0.5 - 1) calc(0.5) / calc(-0.5 + 1))`, `color(${colorSpace} 1.5 -0.5 0.5 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} calc(50% * 3) calc(-150% / 3) calc(50%) / calc(-50% * 3))`, `color(${colorSpace} 1.5 -0.5 0.5 / 0)`);

    test_computed_value("color", `color(${colorSpace} none none none / none)`, `color(${colorSpace} none none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none)`, `color(${colorSpace} none none none)`);
    test_computed_value("color", `color(${colorSpace} 10% none none / none)`, `color(${colorSpace} 0.1 none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none / 0.5)`, `color(${colorSpace} none none none / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / none)`, `color(${colorSpace} 0 0 0 / none)`);
}

for (const colorSpace of [ "xyz", "xyz-d50", "xyz-d65" ]) {
    const resultColorSpace = colorSpace == "xyz" ? "xyz-d65" : colorSpace;

    test_computed_value("color", `color(${colorSpace} 0 0 0)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 1)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1 1)`, `color(${resultColorSpace} 1 1 1)`);
    test_computed_value("color", `color(${colorSpace} 1 1 1 / 1)`, `color(${resultColorSpace} 1 1 1)`);
    test_computed_value("color", `color(${colorSpace} -1 -1 -1)`, `color(${resultColorSpace} -1 -1 -1)`);
    test_computed_value("color", `color(${colorSpace} 0.1 0.1 0.1)`, `color(${resultColorSpace} 0.1 0.1 0.1)`);
    test_computed_value("color", `color(${colorSpace} 10 10 10)`, `color(${resultColorSpace} 10 10 10)`);
    test_computed_value("color", `color(${colorSpace} .2 .2 .25)`, `color(${resultColorSpace} 0.2 0.2 0.25)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 0.5)`, `color(${resultColorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} .20 0 10/0.5)`, `color(${resultColorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} .20 0 10/50%)`, `color(${resultColorSpace} 0.2 0 10 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / -10%)`, `color(${resultColorSpace} 0 0 0 / 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 110%)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / 300%)`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1)`, `color(${resultColorSpace} 1 1 0)`);
    test_computed_value("color", `color(${colorSpace} 1)`, `color(${resultColorSpace} 1 0 0)`);
    test_computed_value("color", `color(${colorSpace})`, `color(${resultColorSpace} 0 0 0)`);
    test_computed_value("color", `color(${colorSpace} 1 1 / .5)`, `color(${resultColorSpace} 1 1 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 1 / 0.5)`, `color(${resultColorSpace} 1 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} / 50%)`, `color(${resultColorSpace} 0 0 0 / 0.5)`);
    test_computed_value("color", `color(${colorSpace} calc(0.5 + 1) calc(0.5 - 1) calc(0.5) / calc(-0.5 + 1))`, `color(${resultColorSpace} 1.5 -0.5 0.5 / 0.5)`);

    test_computed_value("color", `color(${colorSpace} none none none / none)`, `color(${resultColorSpace} none none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none)`, `color(${resultColorSpace} none none none)`);
    test_computed_value("color", `color(${colorSpace} 0.2 none none / none)`, `color(${resultColorSpace} 0.2 none none / none)`);
    test_computed_value("color", `color(${colorSpace} none none none / 0.5)`, `color(${resultColorSpace} none none none / 0.5)`);
    test_computed_value("color", `color(${colorSpace} 0 0 0 / none)`, `color(${resultColorSpace} 0 0 0 / none)`);
}

for (const colorSpace of [ "lab", "oklab" ]) {
    test_computed_value("color", `${colorSpace}(0% 0 0)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / 1)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / 0.5)`, `${colorSpace}(0% 0 0 / 0.5)`);
    test_computed_value("color", `${colorSpace}(20% 0 10/0.5)`, `${colorSpace}(20% 0 10 / 0.5)`);
    test_computed_value("color", `${colorSpace}(20% 0 10/50%)`, `${colorSpace}(20% 0 10 / 0.5)`);
    test_computed_value("color", `${colorSpace}(400% 0 10/50%)`, `${colorSpace}(400% 0 10 / 0.5)`);
    test_computed_value("color", `${colorSpace}(50% -160 160)`, `${colorSpace}(50% -160 160)`);
    test_computed_value("color", `${colorSpace}(50% -200 200)`, `${colorSpace}(50% -200 200)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / -10%)`, `${colorSpace}(0% 0 0 / 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / 110%)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / 300%)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(-40% 0 0)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(50% -20 0)`, `${colorSpace}(50% -20 0)`);
    test_computed_value("color", `${colorSpace}(50% 0 -20)`, `${colorSpace}(50% 0 -20)`);
    test_computed_value("color", `${colorSpace}(calc(50% * 3) calc(0.5 - 1) calc(1.5) / calc(-0.5 + 1))`, `${colorSpace}(150% -0.5 1.5 / 0.5)`);
    test_computed_value("color", `${colorSpace}(calc(-50% * 3) calc(0.5 + 1) calc(-1.5) / calc(-0.5 * 2))`, `${colorSpace}(0% 1.5 -1.5 / 0)`);

    test_computed_value("color", `${colorSpace}(none none none / none)`, `${colorSpace}(none none none / none)`);
    test_computed_value("color", `${colorSpace}(none none none)`, `${colorSpace}(none none none)`);
    test_computed_value("color", `${colorSpace}(20% none none / none)`, `${colorSpace}(20% none none / none)`);
    test_computed_value("color", `${colorSpace}(none none none / 0.5)`, `${colorSpace}(none none none / 0.5)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / none)`, `${colorSpace}(0% 0 0 / none)`);
}

for (const colorSpace of [ "lch", "oklch" ]) {
    test_computed_value("color", `${colorSpace}(0% 0 0deg)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0deg / 1)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0deg / 0.5)`, `${colorSpace}(0% 0 0 / 0.5)`);
    test_computed_value("color", `${colorSpace}(100% 230 0deg / 0.5)`, `${colorSpace}(100% 230 0 / 0.5)`);
    test_computed_value("color", `${colorSpace}(20% 50 20deg/0.5)`, `${colorSpace}(20% 50 20 / 0.5)`);
    test_computed_value("color", `${colorSpace}(20% 50 20deg/50%)`, `${colorSpace}(20% 50 20 / 0.5)`);
    test_computed_value("color", `${colorSpace}(10% 20 20deg / -10%)`, `${colorSpace}(10% 20 20 / 0)`);
    test_computed_value("color", `${colorSpace}(10% 20 20deg / 110%)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(10% 20 1.28rad)`, `${colorSpace}(10% 20 73.3386)`);
    test_computed_value("color", `${colorSpace}(10% 20 380deg)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(10% 20 -340deg)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(10% 20 740deg)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(10% 20 -700deg)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(-40% 0 0)`, `${colorSpace}(0% 0 0)`);
    test_computed_value("color", `${colorSpace}(20% -20 0)`, `${colorSpace}(20% 0 0)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / 0.5)`, `${colorSpace}(0% 0 0 / 0.5)`);
    test_computed_value("color", `${colorSpace}(10% 20 20 / 110%)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(10% 20 -700)`, `${colorSpace}(10% 20 20)`);
    test_computed_value("color", `${colorSpace}(calc(50% * 3) calc(0.5 - 1) calc(20deg * 2) / calc(-0.5 + 1))`, `${colorSpace}(150% 0 40 / 0.5)`);
    test_computed_value("color", `${colorSpace}(calc(-50% * 3) calc(0.5 + 1) calc(-20deg * 2) / calc(-0.5 * 2))`, `${colorSpace}(0% 1.5 320 / 0)`);

    test_computed_value("color", `${colorSpace}(none none none / none)`, `${colorSpace}(none none none / none)`);
    test_computed_value("color", `${colorSpace}(none none none)`, `${colorSpace}(none none none)`);
    test_computed_value("color", `${colorSpace}(20% none none / none)`, `${colorSpace}(20% none none / none)`);
    test_computed_value("color", `${colorSpace}(none none none / 0.5)`, `${colorSpace}(none none none / 0.5)`);
    test_computed_value("color", `${colorSpace}(0% 0 0 / none)`, `${colorSpace}(0% 0 0 / none)`);
}
</script>
</body>
</html>
